added atom mapping to cp2k
